Remove fast mode
Ansible does not expose an easy way to manipulate lists in the way
required for fast mode. Since the lists could be dicts or strings and
we would want to transform them into strings we need to map and apply
something like "item.get('name', item)" to them and Ansible AFAICT has
no easy Jinja powered way to do that.

It was never really clear if it would make a big perf diff anyway so
just remove it.