Speed up package installation in the most common case
The package lists are now exactly that, you can opt for a slower but
more finely controlled version of package installation using the
`$format_package_specs` variables. This effectively means that
`*_specs` variables behave as the original package variables and the
original package variable names are now assumed to be lists of
strings.